chore(cli): Merge the utils and util modules
