+++ lib/cli/src/utils/package_wizard.rs
