+++ lib/cli/src/utils/render.rs
