+++ lib/cli/src/utils/templates/js-worker/index.js
