+++ lib/cli/src/utils/templates/py-application/main.py
