+++ lib/cli/src/utils/templates/static-site/index.html
